Weights our now double, not short.

+++ cgworkshop/src/GraphHandler.cpp
@@ -38,26 +38,25 @@ void GraphHandler::init_graph(int rows, int cols, CvMat * smoothness) {
 	{
 		for (int j = 1; j < cols; j += 2) 
 		{
-			// Casting to short here is very very bad.
-			m_igraph->add_edge(m_nodes[i][j], m_nodes[i-1][j-1], (short) calcDist(smoothness, i*cols+j, (i-1)*cols+(j-1)));
-			m_igraph->add_edge(m_nodes[i][j], m_nodes[i][j-1], (short) calcDist(smoothness, i*cols+j, i*cols+(j-1)));
+			m_igraph->add_edge(m_nodes[i][j], m_nodes[i-1][j-1], calcDist(smoothness, i*cols+j, (i-1)*cols+(j-1)));
+			m_igraph->add_edge(m_nodes[i][j], m_nodes[i][j-1], calcDist(smoothness, i*cols+j, i*cols+(j-1)));
 
-			m_igraph->add_edge(m_nodes[i][j], m_nodes[i-1][j], (short) calcDist(smoothness, i*cols+j, (i-1)*cols+j));
+			m_igraph->add_edge(m_nodes[i][j], m_nodes[i-1][j], calcDist(smoothness, i*cols+j, (i-1)*cols+j));
 			
 			if (j < cols - 1)
 			{
-				m_igraph->add_edge(m_nodes[i][j], m_nodes[i-1][j+1], (short) calcDist(smoothness, i*cols +j, (i-1)*cols +(j+1)));		
-				m_igraph->add_edge(m_nodes[i][j], m_nodes[i][j+1], (short) calcDist(smoothness, i*cols + j, i*cols + (j+1)));						
+				m_igraph->add_edge(m_nodes[i][j], m_nodes[i-1][j+1], calcDist(smoothness, i*cols +j, (i-1)*cols +(j+1)));		
+				m_igraph->add_edge(m_nodes[i][j], m_nodes[i][j+1], calcDist(smoothness, i*cols + j, i*cols + (j+1)));						
 			}
 
 			if (i < rows - 1)
 			{
-				m_igraph->add_edge(m_nodes[i][j], m_nodes[i+1][j-1], (short) calcDist(smoothness, i*cols+j, (i+1)*cols+(j-1)));						
-				m_igraph->add_edge(m_nodes[i][j], m_nodes[i+1][j], (short) calcDist(smoothness, i*cols+j, (i+1)*cols+j));	
+				m_igraph->add_edge(m_nodes[i][j], m_nodes[i+1][j-1], calcDist(smoothness, i*cols+j, (i+1)*cols+(j-1)));						
+				m_igraph->add_edge(m_nodes[i][j], m_nodes[i+1][j], calcDist(smoothness, i*cols+j, (i+1)*cols+j));	
 			}
 
 			if ((j < cols - 1) && (i < rows - 1))
-				m_igraph->add_edge(m_nodes[i][j], m_nodes[i+1][j+1], (short) calcDist(smoothness, i*cols+j, (i+1)*cols+(j+1)));				
+				m_igraph->add_edge(m_nodes[i][j], m_nodes[i+1][j+1], calcDist(smoothness, i*cols+j, (i+1)*cols+(j+1)));				
 		}
 	}
 
@@ -71,10 +70,9 @@ void GraphHandler::assign_weights(CvMat *Tu, CvMat *Su) {
 		for (int j=0; j<Tu->cols; j++) {
 			//add the Sink E1 term
 			m_igraph->set_tweights(m_nodes[i][j], 
-										(short)cvmGet(Su,i,j),
-										(short)cvmGet(Tu,i,j));
-										
-		// Casting to short here is very very bad.								
+										cvmGet(Su,i,j),
+										cvmGet(Tu,i,j));
+														
 	}
